+++ pyroller/__init__.py
@@ -1,2 +1,4 @@
+import importlib.metadata
+
 __all__ = ["__version__"]
-__version__ = "0.5.6"
+__version__ = importlib.metadata.version("py-roller")